requirements.txt
tensorflow>=2.11.0
opencv-python
Pillow
numpy
pandas
scikit-learn
matplotlib
flask